# Conflicts:
#	Final/__pycache__/DjangoInterface.cpython-37.pyc
#	Final/__pycache__/views.cpython-37.pyc
#	__pycache__/App.cpython-37.pyc
#	__pycache__/Login.cpython-37.pyc
#	__pycache__/UserEdits.cpython-37.pyc
#	db.sqlite3

+++ UserEdits.py
@@ -1,33 +1,72 @@
 from Final import DjangoInterface
+from Final.models import *
+
+
 class UserEdits:
 
     def __init__(self):
         self = self
-#        self.myDj = DjangoInterface()
 
-    def add_user(self, username, password, loggedinuser):
-        print("User in Edits: " + username)
-        print("Pass in Edits: " + password)
+    def add_user(self, username, password, permissions, logged_in_user):
+        if logged_in_user.permissions[0] != '1' and logged_in_user.permissions[1] != '1':
+            return "Illegal permissions to do this action"
+        if len(permissions) != 4:
+            return "Failed to add user. Improper parameters"
+        if not username.strip():  # this equates to if usersame == " "
+            return "Failed to add user. Improper parameters"
+        if not password.strip():
+            return "Failed to add user. Improper parameters"
+        if not permissions.strip():
+            return "Failed to add user. Improper parameters"
+        if '*' in username:
+            return "Failed to add user. Improper parameters"
+        if len(password) < 2:
+            return "Failed to add user. Improper parameters"
         try:
-            DjangoInterface.DjangoInterface.create_user(self, username, password)
+            DjangoInterface.DjangoInterface.create_user(self, username, password, permissions)
         except Exception as e:
             print(e)
-            return "Error"
-        return "User successfully added"  #Whatever was written in the acceptance tests
+            return "Failed to add user."
+        return "User successfully added"  # Whatever was written in the acceptance tests
 
-    def delete_user(self, usertodelete, loggedinuser):
+    def delete_user(self, usertodelete, logged_in_user):
+        if logged_in_user.permissions[0] != '1' and logged_in_user.permissions[1] != '1':
+            return "Illegal permissions to do this action"
+        if usertodelete == logged_in_user.username:
+            return "Unable to delete active user"
         try:
             DjangoInterface.DjangoInterface.delete_user(self, usertodelete)
         except Exception as e:
-            return "Error"
+            return "User unsuccessfully deleted"
         return "User successfully deleted"
 
-    def change_contact(self, usernameToChange, fieldToChange, updatedInfo, loggedInUser):
-        try:
-            DjangoInterface.DjangoInterface.update_user(self, usernameToChange, fieldToChange, updatedInfo)
-        except Exception as e:
-            return "Error"
-        return "User updated"
+    def change_contact(self, user_to_change, field_to_change, updated_field):
+        if field_to_change == "username":
+            return "Invalid parameter for this command"
+        all_field_names = User._meta.get_fields()
+        for field in all_field_names:
+            if field_to_change == field.attname:
+                try:
+                    DjangoInterface.DjangoInterface.update_user(self, user_to_change, field_to_change, updated_field)
+                except Exception as e:
+                    print(e)
+                    return "Failed to update user"
+                return "Contact information changed"  # Whatever was written in the acceptance tests
+        return "Illegal changed field"
 
-    def edit_user(self, usertoedit, desiredfield, changedfield, loggedinuser):
-        return "yes"
+    def edit_user(self, user_to_edit, field_to_change, updated_field, logged_in_user):
+        if logged_in_user.permissions[0] != '1' and logged_in_user.permissions[1] != '1':
+            return "Illegal permissions to do this action"
+        if field_to_change == "username" and '*' in updated_field:
+            return "Failed to updated user"
+        all_field_names = User._meta.get_fields()
+        for field in all_field_names:
+            if field_to_change == field.attname:
+                try:
+                    DjangoInterface.DjangoInterface.update_user(self, user_to_edit, field_to_change, updated_field)
+                except Exception as e:
+                    print(e)
+                    return "Failed to update user"
+                return "User successfully updated"  # Whatever was written in the acceptance tests
+        return "Tried to change illegal field"
+
